yazi duzenlemeleri

+++ 6.giris_cikis_fonsiyonlari.c
@@ -1,5 +1,5 @@
-        #include<stdio.h>
-        #define PI 3.14
+     #include<stdio.h>
+    #define PI 3.14
         // giriþ fonksiyon scanf 
         //çýkýþ fonksiyon print 
     
@@ -9,47 +9,17 @@
          float olan bir x tanýmlayýp
          daha sonra bu deðeri kullanýcýdan almak istersek de scanf("%f",&x); þeklinde yazýyoruz. "&" iþareti x'e kullanýcýdan gelen deðeri atadýðýmýzý gösteriyor. 
          Diðer format belirleyicileri printf ile aynýdýr.
-		
-		
-		
-		
-		
-		
 		*/
-
-
          int main() {
-         	
-         	
-         	
-        int yaricap;
+        float yaricap;
         float hacim;
         
         printf("kurenin yari capini giriniz:");
-        scanf("%d",&yaricap);
-        hacim = (4/3.0)*PI*(yaricap*yaricap*yaricap);
+        
+        scanf("%d",& yaricap);
+        
+        hacim = (4 / 3.0)*PI*(yaricap*yaricap*yaricap);
         printf("Kurenýn hacmi %2.f 'dir",hacim);
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	
-         	return 0;
-		 }
+        
+       	return 0;
+        }
